MGMT-6445: Introduce ClusterDeploymentWizard and related refactoring
