+++ src/components/ClusterDeployment/ClusterDeploymentWizard.tsx
@@ -1,0 +1,35 @@
+import * as React from 'react';
+import { AlertsContextProvider } from '../AlertsContextProvider';
+import { ClusterDeploymentWizardProps, ClusterDeploymentWizardStepsType } from './types';
+import ClusterDeploymentDetails from './ClusterDeploymentDetails';
+import ClusterDeploymentWizardNavigation from './ClusterDeploymentWizardNavigation';
+
+const ClusterDeploymentWizard: React.FC<ClusterDeploymentWizardProps> = (props) => {
+  const [currentStepId, setCurrentStepId] = React.useState<ClusterDeploymentWizardStepsType>(
+    'cluster-details',
+  );
+
+  const moveNext = () => setCurrentStepId('todo-next-wizard-step-id');
+  const navigation = (
+    <ClusterDeploymentWizardNavigation
+      currentStepId={currentStepId}
+      setCurrentStepId={setCurrentStepId}
+    />
+  );
+
+  const renderCurrentStep = () => {
+    switch (currentStepId) {
+      case 'cluster-details':
+      default:
+        return <ClusterDeploymentDetails moveNext={moveNext} navigation={navigation} {...props} />;
+    }
+  };
+
+  return (
+    <AlertsContextProvider>
+      <div className="pf-c-wizard">{renderCurrentStep()}</div>;
+    </AlertsContextProvider>
+  );
+};
+
+export default ClusterDeploymentWizard;
