+++ src/components/ClusterDeployment/index.ts
@@ -1,0 +1,2 @@
+export * from './types';
+export { default as ClusterDeploymentWizard } from './ClusterDeploymentWizard';
