+++ src/components/clusterWizard/index.ts
@@ -1,0 +1,1 @@
+export { default as ClusterDetailsForm } from './ClusterDetailsForm';
